Update reservation and add connection on code validation.
