+++ src/cloud/validateCode.js
@@ -4,12 +4,13 @@ import generatePassword from '../utils/generatePassword';
 import TwoFAService from '../services/TwoFAService';
 import UserService from '../services/UserService';
 import ReservationService from '../services/ReservationService';
+import ConnectionService from '../services/ConnectionService';
 
 class ValidateCodeError extends ExtendableError {}
 
 const validateCode = async request => {
   let { params, installationId } = request;
-  const { phoneNumber, authCode } = params;
+  const { phoneNumber, authCode, reservationId } = params;
 
   // Phone number is required in request body
   if (!phoneNumber) {
@@ -52,6 +53,19 @@ const validateCode = async request => {
       user.set('verificationStatus', status);
       user.set('verificationValid', valid);
       await user.save(null, { useMasterKey: true });
+
+      if (reservationId) {
+        const reservation = await ReservationService.checkReservation(
+          reservationId,
+        );
+
+        // set reservation as claimed and create a connection between users
+        reservation.set('isClaimed', true);
+        reservation.set('user', user);
+        await reservation.save(null, { useMasterKey: true });
+        const fromUser = reservation.get('createdBy');
+        await ConnectionService.createConnection(fromUser, user);
+      }
       // creates 3 reservations for the new user.
       // TODO: set this number as an app configuration.
       await ReservationService.createReservations(user, 3);
